Fix bug in cycle detection
